Improve authorization
